Attention is All You Need (2017)
The 2017 paper “Attention is All You Need” done mostly by Google guys conceived the architecture of LLMs and Transformer. OpenAI just productized it. The release of ChatGPT in Nov 2022 was controversial because a. it’s not OpenAI’s idea; b. OpenAI became a private company; c. most AI inner circle believed the world was not ready for GPT-3.