Statistical vs. Logical
Many people are confused abt the difference between statistical vs logical. LLMs are statistical & inherently consensus based. Logic, maths, & reasoning don’t need consensus to be right. They are right no matter how small the portion of the data represent it. This is the limitation of the LLM architecture. It ain’t the path to AGI.